CRUD 정리 CoreDataStack정리

+++ FoodTracker/Meal.swift
@@ -14,9 +14,9 @@ class Meal{
     //MARK: Properties
     var name: String?
     var photo: UIImage?
-    var rating: Int?
-    
+    var rating: Int16?
+
     //원본 MealMO 객체를 참조하기 위한 속성
     var objectID: NSManagedObjectID?
-   
+
 }
